// src/Components/Text/type.ts
/*
 * File Created: Saturday, 31st July 2021 3:58:28 pm
 * Author: Bruno Oliveira (bruno.oliveira@zappts.com)
 * -----
 * Last Modified: Saturday, 31st July 2021 4:11:52 pm
 * Modified By: Bruno Oliveira (bruno.oliveira@zappts.com)
 * -----
 * Copyright 2021 - Cartões ADD, Porto Seguro
 */

export default interface TextProps {
  size?: number;
  color?: string;
  uppercase?: boolean;
  fontFamily?: string;
  marginLeft?: number;
  marginTop?: number;
  fontWeight?: number;
  marginRight?: number;
  marginBottom?: number;
  textAlign?: string;
  textDecorationColor?: string;
}